+++ src/activeTab/formatter/formatText.js
